Add startNewRound() button
Created a button with the trigger onclick in index.html attached to a button 'play'. This was done to make the game easier to play, as well as test. Button highlighting was removed. Instructions were also added/altered under the index.html file to make it reflect the changes, this will hopefully make the game more user friendly when complete.