Move console.log() stats
Moved the console.log of game stats to roundCounter. This will not print the stats of the game every single round instead of on completion of a match. This was done to see why some rounds would end with a draw when player or computer reached 3 wins. Will leave so player can see what score is at each round