Add error handling to playgame()
Added an else if statement and else statement to playGame. Chrome has a violation warning which occurs when the function takes too long to play, this stops the game from counting correctly. Else if takes the input of start, this will be added to index.html later.

Else statement added to catch input in prompt which are not part of the game. If a player enters an invalid string, the game will remove round from count and run startNewRound(). This prevents the game from crashing on an invalid string.