Add new variables
Started work on adding a GUI for the app. Added new container in index.html, a coresponding style rule was created to set display grid. Finally created 3 new const variables within the js file. This is the foundation for adding a GUI way to play the game instead of using a terminal to play. I will leave the terminal method as well; however, add the GUI for most players