Remove redundancy in set functions
Removed setRock(), setPaper(), and setScissor() functions. Also removed computerPlay() function. This was done to reduce redundant code. Buttons now call playGame() with their respective values being called as arguments. computerPlay() was removed and the variable is now declared and assigned in playGame(). This reduces the need for a seperate function.