Add function getGridSize()
Added a function which takes input from an text input node. Submit button calls setGridSize() which checks if the number is between 1 and 64, also if the value is of the number object type. If the user inputs an invalid option an alert is triggered and the grid size will be set to the default 16x16 grid